Brittle attempt to cleanup empty files, that actually works
