+++ test/shared/src/main/scala/zio/test/TestExecutor.scala
@@ -162,7 +162,7 @@ object TestExecutor {
             )
             ZIO.scoped {
               loop(List.empty, scopedSpec, defExec, List.empty, topParent)
-            } *> processEvent(topLevelFlush) *> deleteIfEmpty("target/test-reports-zio/last_executing.txt")
+            } *> processEvent(topLevelFlush)
           }
           summary <- summary.get
         } yield summary).provideLayer(sinkLayer)
@@ -173,23 +173,6 @@ object TestExecutor {
           case Right(testSuccess) => testSuccess.annotations
         }
 
-      private def deleteIfEmpty(path: String)(implicit trace: Trace) = ZIO.succeed{
-        import java.io._
-        import scala.io.Source
-
-        val lines = Source.fromFile(path).getLines.filterNot(_.isBlank).toList
-        val file = new File(path)
-        if (file.exists() && lines.isEmpty) {
-          if (file.delete()) {
-            println("File deleted successfully.")
-          } else {
-            println("Failed to delete the file.")
-          }
-        } else {
-          println("File does not exist.")
-        }
-
-      }
 
     }
